In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,15 +10,15 @@
  */
 int main(void)
 {
-unsigned int a = 1, b = 2, next;
+unsigned long long int a = 1, b = 2, next;
 int count;
 
-printf("%u, %u, ", a, b);
+printf("%llu, %llu, ", a, b);
 
 for (count = 3; count <= 98; count++)
 {
 next = a + b;
-printf("%u", next);
+printf("%llu", next);
 
 if (count < 98)
 printf(", ");
